refactor: move streaming helpers to agent_runner module
